Get Single Community Group, Add Community Group and route renaming
