+++ src/index.js
@@ -1,5 +1,6 @@
 const express = require("express");
 const morgan = require("morgan");
+const bodyParser = require("body-parser");
 
 const router = require("./api");
 const { logger } = require("./utils/logger");
@@ -13,6 +14,7 @@ const port = 3000;
 
 logger.info("🤖 Initializing middleware");
 
+app.use(bodyParser.json());
 app.use(morgan("tiny", { stream: logger.stream }));
 app.use("/", router);
 app.use(errorHandler);
