--- src/routes/communityGroups/communityGroups.contoller.js
@@ -1,16 +1,0 @@
-const data = require("../../../db/communityGtoups.data.json");
-
-const listCommunityGroups = (req, res) => {
-  return res.json(data);
-};
-
-const getCommunityGroup = (req, res) => {
-  console.log(req.params.id);
-  // const record = data.filter(record => record.id === req.params.id);
-  // console.log(record);
-};
-
-module.exports = {
-  listCommunityGroups,
-  getCommunityGroup
-};
