+++ src/routes/communityGroups/community_groups.router.js
@@ -2,13 +2,15 @@ const express = require("express");
 
 const {
   listCommunityGroups,
-  getCommunityGroup
-} = require("./communityGroups.contoller");
+  getCommunityGroup,
+  addCommunityGroup
+} = require("./community_groups.contoller");
 
 const router = express.Router();
 
 router.get("", listCommunityGroups);
 router.get("/:id", getCommunityGroup);
+router.post("", addCommunityGroup);
 
 module.exports = {
   communityGroups: router
